Fixed include paths in documentation
I noticed that these are pointing to a perhaps older directory structure, as the include paths were incorrect in the documentation of Namespace Configuration in the README.md from how the document structure currently is attainable via Composer. I have a suspicion that the Yaml include path may be different, but I do not currently use this component.